==> notebooks/02_Inicjowanie_bazy_danych.ipynb <==
 # Opis notatnika
 W tym notatniku skupimy się na stworzeniu dedykowanej bazy danych wraz ze strukturą tabel. Jednak w tym momencie nie będziemy jej uzupełniać danymi - tym zajmiemy się w kolejnej części.
 
 Ponownie wcielasz się w rolę Data Engineera, którego zadaniem jest dostarczenie klientowi bazy danych, która będzie używana do celów analityczno-raportowych. Bazy danych są podstawowym źródłem pracy dla Analityków Danych.
 
 Plik wraz ze strukturą bazy znajdziesz w folderze `sql/database_schema.sql`. 

 > W tym zadaniu konieczne jest posiadanie zainstalowanego `Postgres` na swoim komputrze. Jeżeli z jakichś powodów został on usunięty po kursie `Python - Analiza Danych`, należy ponownie dokonać instalacji.

 > Posiadając działającą instancję serwera, można stworzyć po prostu nową bazę np. `airlines`. Stworzenie bazy danych powinno odbyć się w notatniku poprzez stworzenie obiektu cursor i wywołaniu zapytań znajdujących się w pliku .sql.

 W trakcie rozwiązywania tego notatnika przydadzą się następujące materiały:
 - `Python - analiza danych -> Przygotowanie do kursu -> Podstawy SQL - Praca samodzielna -> Instalacja bazy danych`
 - `Python - analiza danych -> Moduł 1 -> Dzień 3 - PostgreSQL -> SQL i Python`
 - `Python - analiza danych -> Prework -> Podstawy SQL - Praca samodzielna -> Instalacja bazy danych -> Przygotowanie bazy danych`
 
Inicjację bazy danych wykonaj w notatniku, nie bezpośrednio na Postgresie.

 Tutaj zaimportuj potrzebne biblioteki

from psycopg2 import connect, extensions

 ## Połączenie z bazą danych
 Tutaj uzupełnij konfigurację połączenia

username = 'postgres'
password = '123'

host = 'localhost'
database = 'airlines'
port = None


Tutaj zdefiniuj zmienną `con` oraz `cursor`

con = connect(
    host=host,
    user=username,
    password=password,
    database=database
)

con.set_isolation_level(extensions.ISOLATION_LEVEL_AUTOCOMMIT)

cursor = con.cursor()

cursor.execute("CREATE DATABASE airlines")

con.commit()

 ## Wczytanie pliku `database_schema.sql`
 Z katalogu `sql` wczytaj plik `database_schema.sql`

with open(r'..\sql\database_schema.sql') as sql_file:
    sql_txt = sql_file.read()

 W tym miejscu odpowiednio rozdziel zawartość pliku `database_schema.sql` na mniejsze kwerendy używając `;`

sql_commands = sql_txt.split(';')
print(sql_commands)

 W tym miejscu wykonaj każdą z kwerend, aby zainicjować strukturę bazy danych

for command in sql_commands:
    cursor.execute(command+';')

 Zatwierdznie wszystkich operacji na bazie, czyli stworzenie tabel

con.commit()

 ### Sprawdzenie
 Uruchom kod poniżej, aby sprawdzić, czy ta część została poprawnie wykonana

def check_if_table_exists(table_name):
    msg = f"Sprawdzam czy istnieje tabela {table_name}"
    print(msg)

    query = f"select 1 from {table_name}"
    # jeżeli tabela nie istnieje, ten krok zwróci wyjątek
    cursor.execute(query)
    print('OK!')


tables_to_test = [
    'aircraft',
    'airport_weather',
    'flight',
    'airport_list'
]

for table in tables_to_test:
    check_if_table_exists(table)

con.close()
msg = "Wszystko wygląda OK :) Możesz przejść do kolejnego zadania."
print(msg)

 # Podsumowanie
 Za pomocą tego notatnika została zbudowana w sposób automatyczny nasza docelowa baza danych. Dzięki temu nie musimy się już martwić o jej ręczną przebudowę - w ramach potrzeby wystarczy włączyć notatnik.
 Wykonując kolejny notatnik, sprawimy, że w tabelach pojawią się również dane potrzebne do przeprowadzenia późniejszej analizy oraz wykorzystywane w raportowaniu.
