==> notebooks/06_Dashboard.ipynb <==
 # Opis zadania
 W tym notatniku stworzymy system analityczno-raportowy, mający wspomagać przyszłe decyzje biznesowe. Dzięki przeprowadzonej wcześniej analizie eksploracyjnej danych, w trakcie spotkania prezentującego wykonaną pracę postanowiono w pierwszym kroku stworzyć następujące wizualizacje danych:
 - Wyświetlenie TOP 10 tras samolotów pod względem najniższego odsetka opóźnień w formie tabeli,
 - Wyświetlenie porównania roku 2019 oraz 2020 w formie wykresu słupkowego:
     - miesiąc do miesiąca,
     - dzień tygodnia do dnia tygodnia,
 - Wyświetlenie danych dzień po dniu w formie szeregu czasowego.

Raport powinien składać się z trzech stron - `TOP report`, `Comparision` i `Day by day reliability`.

Pracując na podstawie wcześniej zdefiniowanych widoków, nie musimy się tutaj martwić o przetwarzanie oraz procesowanie danych. Co więcej, wszystkie niuanse techniczne w postaci liczby wierszy, wykonywanych złączeń, filtracji są przeniesione na bazę danych. Z technicznego punktu widzenia, pomiędzy serwisem a bazą danych dochodzi do przesyłania mniejszej ilości danych.

Dopuszczamy jednak pewne aktualizacje co do ich struktury przykładowo poprzez _pivotowanie_, czyli obranie i doprowadzanie do postaci tej znanej z tabel przestawnych.

> **Uwaga:**  
> Przy pracy nad poprawkami w dashboardzie, pamiętaj, że aby odświeżyć stronę po wprowadzonych zmianach, należy **całość** uruchomić ponownie.

W tej części projektu końcowego wcielasz się ponownie w rolę BI Developera, który ma za zadanie stworzyć dashboard zgodny z wytycznymi biznesowymi dostarczonymi przez klienta. Pamiętaj, że osoba na tym stanowisku często ma kontakt z biznesem więc musi umieć przekazać informację o danych, które zawarła na dashboardzie by te zaangażowały odbiorców.


 Tutaj zaimportuj potrzebne biblioteki

import dash
from dash import dcc
from dash import html
from dash.dependencies import Input, Output
from dash import dash_table

import pandas as pd
import psycopg2
from sqlalchemy import create_engine
from sqlalchemy.engine import URL
import plotly.express as px

 ## Konfiguracja połączenia
 Tutaj uzupełnij konfigurację połączenia

username = 'postgres'
password = '123'

host = 'localhost'
database = 'airlines'
port = None

url = URL.create(
    "postgresql",
    username = 'postgres',
    password = '123',
    host = 'localhost',
    database = 'airlines')
    
engine = create_engine(url)

 Tutaj zdefiniuj zmienną `con` oraz `cursor`

# con = connect(
#     host=host,
#     user=username,
#     password=password,
#     database=database
# )

# cursor = con.cursor()

 # Stworzenie tabeli dla `TOP 10 routes`
 W tym miejscu przygotujemy tabelkę oraz komponent, który zostanie następnie umieszczony w raporcie. Tabela powinna się składać z następujących kolumn:
 - Origin,
 - Destination,
 - Year,
 - Reliability,
 - Rank.

Wartości procentowe zaś powinny być postaci np. 13.87%.

Z tego względu konieczna będzie zmiana nazw kolumn dla wyników raportowania.

Wskazówki:
 - `Python - analiza danych > Dzień 5 - Pandas > Obróbka danych - część 2 > apply`,
 - [How to format percentage in python](https://stackoverflow.com/questions/5306756/how-to-print-a-percentage-value-in-python).

 Tutaj zaczytaj dane do ramki danych `top_routes_df` z widoku `reporting.top_reliability_roads`
 > Jeśli pojawi się komunikat `UserWarning: pandas only support SQLAlchemy`, z naszego punktu widzenia, możemy go śmiało zignorować.

query = "SELECT * FROM reporting.top_reliability_roads;"
top_routes_df = pd.read_sql(query, engine)

top_routes_df

 Tutaj dokonaj aktualizacji formatu liczbowego dla kolumny `reliability`

# w budowie schematu proszono o stworzenie widoku i rankingu według najwyższego odsetka opóźnień
top_routes_df['delayed_ratio_reliability'] = (top_routes_df['delayed_ratio_reliability'] * 100).astype(str) + '%'
top_routes_df

 Tutaj odpowiednio dokonaj zmian nazewnictwa kolumn

top_routes_df = top_routes_df.loc[:, ['origin_airport_name', 'dest_airport_name', 'year', 'delayed_ratio_reliability', 'nb']].rename(columns={
    'origin_airport_name': 'Origin Airport',
    'dest_airport_name': 'Destination Airport',
    'year': 'Year',
    'delayed_ratio_reliability': 'Not reliable',
    'nb': 'Rank'
})
top_routes_df

 # Porównanie rok do roku (YoY) 2019 vs. 2020
 W tym miejscu stworzymy wykres oraz komponent, który zostanie następnie umieszczony w raporcie. Wykres powinien przedstawiać porównanie lat 2019 oraz 2020 po miesiącach względem zmiennej `reliability`.

 > Na wykresie chcemy również przedstawić dane, które nie mają porównania tj. od kwietnia do grudnia 2019.

W tym celu wykonamy poniższe zadania:
 - Przekształcimy tabelkę do formy tabeli przestawnej
 - Stworzymy wykres słupkowy porównujący dane rok do roku
 - Opakujemy stworzony wykres w komponent `Dash`

Wskazówki:
 - `Wizualizacja danych > Dzień 3 > Dash`,
 - `Wizualizacja danych > Dzień 1 > Budowanie wykresów plotly`,
 - `Python - analiza danych > Dzień 6 - Pandas c.d > Pivot`.

 Tutaj pobierz z widoku `reporting.year_to_year_comparision` dane do ramki `yoy_comparision_df`

query = "SELECT * FROM reporting.year_to_year_comparision;"
yoy_comparison_df = pd.read_sql(query, engine)
yoy_comparison_df['not_reliable_ratio'] = round(yoy_comparison_df['not_reliable_ratio'] * 100, 2)
yoy_comparison_df

 Tutaj odpowiednio przekształć ramkę, do formy oczekiwanej przez wykres słupkowy. Wyniki zapisz do zmiennej `yoy_comparision_to_plot_df`

yoy_comparison_to_plot_df = yoy_comparison_df.pivot(index='month', columns='year', values='not_reliable_ratio')
yoy_comparison_to_plot_df

 Tutaj stwórz odpowiedni wykres, zapisz do na później pod nazwą `yoy_comparision_fig`. Pamiętaj o dobrych praktykach - niech wykres posiada wszystkie opisy oraz opisowe etykiety osi.

yoy_comparison_fig = px.bar(yoy_comparison_to_plot_df, x=yoy_comparison_to_plot_df.index, y=[2019, 2020],
             title="Percentage of delayed flights by month (2019 vs. 2020)",
             labels={"index": "Month", "value": "Delayed flights (%)", "variable": "Year"},
             barmode='group', # Set barmode to 'group' for grouped bars
             #color_discrete_map={'2019': '#14487A', '2020': '#FAA275'}
             )  

yoy_comparison_fig.update_layout(xaxis_title = 'Month')

month_label = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec']
yoy_comparison_fig.update_xaxes(tickvals=[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12], ticktext=month_label)

yoy_comparison_fig.show()


 # Porównanie dzień tygodnia do dnia tygodnia (WKoWK) 2019 vs. 2020
 Wykonaj analogiczne do poprzedniego kroku, używając jednak w tym momencie danych z widoku `reporting.day_to_day_comparision`

 Tutaj pobierz z widoku `reporting.day_to_day_comparision` dane do ramki `day_to_day_comparision_df`

query = 'SELECT * FROM reporting.day_to_day_comparision'
day_to_day_comparison_df = pd.read_sql(query, engine)
day_to_day_comparison_df['not_reliable_ratio'] = (day_to_day_comparison_df['not_reliable_ratio']) * 100
day_to_day_comparison_df

 Tutaj odpowiednio przekształć ramkę, do formy oczekiwanej przez wykres słupkowy. Wyniki zapisz do zmiennej `day_to_day_comparision_to_plot_df`

day_to_day_comparison_to_plot_df = day_to_day_comparison_df.pivot(index = 'day_of_week', columns = 'year', values = 'not_reliable_ratio')
day_to_day_comparison_to_plot_df

 Tutaj stwórz odpowiedni wykres, zapisz do na później pod nazwą `day_to_day_comparision_fig`. Pamiętaj o dobrych praktykach - niech wykres posiada wszystkie opisy oraz opisowe etykiety osi.

day_to_day_comparison_fig = px.bar(day_to_day_comparison_to_plot_df, x=day_to_day_comparison_to_plot_df.index, y=[2019,2020],
                            title="Percentage of delayed flights by day of week (2019 vs. 2020)",
                            labels={"index": "Day of week", "value": "Delayed flights (%)", "variable": "Year"},
                            barmode='group',
                            #color_discrete_map={'2019': '#14487A', '2020': '#FAA275'}
                            )  

day_to_day_comparison_fig.update_layout(xaxis_title = 'Day of week')

day_label = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun']
day_to_day_comparison_fig.update_xaxes(tickvals=[1, 2, 3, 4, 5, 6, 7], ticktext=day_label)

day_to_day_comparison_fig.show()

 # Stworzenie szeregu czasowego
 W tym miejscu stworzymy wykres w formie szeregu czasowego, który zostanie umieszcony w raporcie. Wykres powinien przedstawiać dane w formie szeregu czasowego (dzień po dniu) dla lat 2019 oraz 2020. Dla możliwości identyfikacji poszczególnych lat, zostaliśmy dodatkowo poproszeni o nadanie innego koloru dla roku 2019 oraz 2020.

 > Pamiętaj o stworzeniu wykresu zgodnie z dobrymi praktykami.

 W tym miejscu pobierz dane do ramki `day_by_day_reliability_df`, z widoku `reporting.day_by_day_reliability`.

query = 'SELECT * FROM reporting.day_by_day_reliability'
day_by_day_reliability_df = pd.read_sql(query, engine)
day_by_day_reliability_df['not_reliable_ratio'] = (day_by_day_reliability_df['not_reliable_ratio']) * 100
day_by_day_reliability_df['year'] = pd.to_datetime(day_by_day_reliability_df['date_column']).dt.year
day_by_day_reliability_df = day_by_day_reliability_df.sort_values(by='date_column')
day_by_day_reliability_df

 Tutaj stwórz wykres liniowy na podstawie pobranych danych. Wynik zapisz do zmiennej `day_by_day_reliability_fig`.

day_by_day_reliability_fig = px.line(day_by_day_reliability_df, x='date_column', y='not_reliable_ratio',
                                    color='year',
                                    title = 'Percentage of delayed flights in 2019 and 2020',
                                    labels={"date_column": "Date", "not_reliable_ratio": "Delayed flights (%)", "year": "Year"},
                                    #color_discrete_map={'2019': '#14487A', '2020': '#FAA275'}
                                    )

day_by_day_reliability_fig.update_xaxes(title_text="Date")
day_by_day_reliability_fig.update_yaxes(title_text="Delayed flights (%)")


day_by_day_reliability_fig.show()

 # Stworzenie layoutów poszczególnych stron

 ## Stworzenie layout dla widoku `TOP 10 routes`
 W tym momencie przekonwertujemy zmienną `top_routes_df` w postać `Dash.DataTable`. Dodatkowo, dla czytelności, dodamy nagłówek opisujący co tabela przedstawia.

 Nagłówek wystylizuj używając poniższych wymagań:
 - tag: `H3`,
 - czcionka: `verdana`,
 - kolor: '#4444`,
 - tekst do wyświetlenia: `TOP 10 reliability routes in 2019 and 2020`

Wskazówki:
 - `Wizualizacja danych > Dzień 3 > Dash`,
 - `Wizualizacja danych > Dzień 3 > Dash - datatable`,
 - Dokmentacja metody `Pandas` - [to_dict](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.to_dict.html).

 Tutaj przygotuj komponent `Dash` `DataTable`, zapisując wynik do zmiennej `top_routes_table`

top_routes_table = dash_table.DataTable(id='datatable',
                                        columns=[{'name': col, 'id': col} for col in top_routes_df.columns],
                                        data=top_routes_df.to_dict('records'),)

 Tutaj przygotuj nagłówek dla tabeli, zapisz go pod nazwą `top_routes_page_title`

top_routes_page_title = html.H3("TOP 10 reliability routes in 2019 and 2020",
                                style={'font-family': 'verdana', 'color': '#4444'}) # 'white'

 Tutaj przygotuj layout dla widoku `TOP 10 routes`. Konfigurację zapisz do zmiennej `top_routes_layout`

top_routes_layout = html.Div([top_routes_page_title, top_routes_table])

 ## Stworzenie layout dla widoku `Comparision`
 W tym momencie opakujemy zmienne `yoy_comparision_fig` oraz `day_to_day_comparision_fig` do komponentów `Dash`, a następnie stworzymy odpowiedni `layout` dla strony `comparision`. Wyniki zapiszemy do zmiennej `comparision_layout`.

 Tutaj opakuj `yoy_comparision_fig` w odpowiedni komponent `Dash`. Wynik zapisz do zmiennej `yoy_comparision_component`.

yoy_comparison_component = dcc.Graph(id='yoy bar', figure=yoy_comparison_fig)

 Tutaj opakuj `day_to_day_comparision_fig` w odpowiedni komponent `Dash`. Wynik zapisz do zmiennej `day_to_day_comparision_component`.

day_to_day_comparison_component = dcc.Graph(id='dtd bar', figure=day_to_day_comparison_fig)

 Tutaj stwórz odpowiedni `layout` dla widoku `comparision`. Wynik zapisz do zmiennej `comparision_layout`.

comparison_layout = html.Div(yoy_comparison_component)

 Tutaj stwórz odpowiedni `layout` dla widoku `day by day`. Wynik zapisz do zmiennej `day_by_day_layout`.

day_by_day_layout = html.Div(day_to_day_comparison_component)

 # Konfigurowanie aplikacji
 Posiadamy już wszystkie niezbędne komponenty do stworzenia oraz uruchomienia aplikacji. Naszym zadaniem w tym miejscu będzie odpowiednie jej skonfigurowanie, aby obsługiwała więcej niż jedną stronę.

 Chcemy, aby możliwe było poruszanie się pomiędzy poniższymi stronami:
 - `TOP report` - strona domyślna,
 - `Comparision` - strona dostępna pod adresem: `comparision_reporting`,
 - `Day by day reporting` - strona dostępna pod adresem: `day_by_day_reporting`

Wskazówki:
 - `Wizualizacja danych > Dzień 4 > Callback context`,
 - `Wizualizacja danych > Dzień 4 > Aplikacja multipage`.
 - Używając komponentu [html.Button](https://community.plotly.com/t/button-with-link/11809), możesz łatwo stworzyć ładniejsze linki do poszczególnych stron.

 Tutaj przygotuj layout aplikacji

app = dash.Dash()

app.layout = html.Div([
  dcc.Location(id='url', refresh=False),

  html.A(html.Button("Top report"), href="/"),
  html.A(html.Button("Comparison"), href="/comparison_reporting"),
  html.A(html.Button("Day by day reporting"), href="/day_by_day_reporting"),
  
  html.Div(id='page-content')
])

layout_page_one = html.Div(
    children=[top_routes_layout])

layout_page_two = html.Div(
  children=[comparison_layout])

layout_page_three = html.Div(
  children=[day_by_day_layout])


 Tutaj stwórz procedurę odpowiadającą za nawigację pomiędzy poszczególnymi stronami

@app.callback(
  Output('page-content', 'children'),
  Input('url', 'pathname')
)
def display_page_content(pathname):
 
  if pathname == "/comparison_reporting":
    return layout_page_two
  elif pathname == "/day_by_day_reporting":
    return layout_page_three
  else:
    return layout_page_one

 Tutaj uruchom aplikację

app.run()

 # Podsumowanie
 W tym notatniku stworzyliśmy zgodnie z wymaganiami raport do dyspozycji osób decyzyjnych. Kod źródłowy może w tym momencie zostać przekazany dalej, do działu IT, który następnie wdroży rozwiązanie na serwer dostępny dla każdej zainteresowanej osoby. W praktyce oznacza to koniec naszej pracy nad tym zadaniem. Choć warto dodać, że często po udostępnieniu raportu pojawiają się dodatkowe wymagania oraz komentarze ze względu na informacje, które są tam zawarte.

 W kolejnym notatniku podsumujemy sobie cały warsztat.
