==> notebooks/07_Podsumowanie.ipynb <==
 # Zakończenie
 Po wykonanym warsztacie, czas na krótkie podsumowanie wykonanej pracy.

 W trakcie zadania wcieliliśmy się w rolę osoby odpowiedzialnej za usprawnienie procesu przepływu danych oraz analityki. Wcześniej, ze względu na używanie zewnętrznego narzędzia firma była zmuszona, aby wszystkich tych czynności dokonywać ręcznie. Dzięki udostępnieniu zewnętrznego serwisu pojawiła się możliwość automatyzacji zadania.

 W trakcie prac wykonaliśmy następujące kroki, które zgrupujemy sobie według odpowiadających im zawodów w branży IT:  
 1. `Data engineer`:  
     - pobranie danych z zewnętrznego serwisu API,  
     - przygotowanie oraz załadowanie danych na bazę danych,
 1. `Data analyst`:  
     - analiza otrzymanych danych,  
 1. `BI Developer`:  
     - stworzenie raportu oraz wizualiacji danych.  

 Mając na uwadze szerokie spektrum warsztatu, w momencie zmiany pracy jesteśmy w stanie lepiej dopasować osobiste preferencje - w zależności od tego, w której części czuliśmy się najbardziej komfortowo. Warto jednak zaznaczyć, że taki podział w projekcie występuje w większych projektach.

 > W fazie początkowej projektu, podobnie jak tutaj, za wszystkie elementy najczęściej jest odpowiedzialna jedna osoba.

 # Podziękowania
 Dziękujemy za udział w kursie `CodersLab - DataLab`. Mamy nadzieję, że nabyte tutaj umiejętności posłużą Ci do zmiany pracy lub usprawnienia swoich codziennych aktywności.

 # Ostatnie zadanie (opcjonalne)
 Czy wiesz, że używając komendy np. `%run "06_Dashboard.ipynb"` możesz uruchomić z jednego notebooka inny? Spróbuj stworzyć nowy notebook, np. `_Main.ipynb`, który uruchomi cały warsztatowy kod.

import os

# List of the files you want to run
notebook_files = [
    "..\notebooks\01_Pobieranie_danych.ipynb",
    "..\notebooks\02_Inicjowanie_bazy_danych.ipynb",
    "..\notebooks\03_Populacja_bazy_danych.ipynb",
    "..\notebooks\04_Analiza_danych_01.ipynb",
    "..\notebooks\04_Analiza_danych_02.ipynb",
    "..\notebooks\04_Analiza_danych_03.ipynb",
    "..\notebooks\05_Budowa_schematu_reporting.ipynb",
    "..\notebooks\06_Dashboard.ipynb"
]

# Create a new Jupyter Notebook file named "_Main.ipynb"
main_notebook_name = "_Main.ipynb"
os.system(f"jupyter nbconvert --to notebook --output {main_notebook_name} --execute ''")

# Open the main notebook and append code to run the other notebooks
with open(main_notebook_name, 'a') as main_notebook:
    for notebook_file in notebook_files:
        main_notebook.write(f"\n# Running {notebook_file}\n")
        main_notebook.write(f"%run {notebook_file}\n")

print(f"Created and configured '{main_notebook_name}' to run the other notebooks.")
